Changed dependencies to rely on @Inject
Removed the @Provides methods for Engine and Car and instead annotated their
constructors with @Inject. By doing this the constructors are automatically
used to create new objects. The only thing that I needed to do was to provide
a @Provides method that provides Integer since the Engine constructor has
an int as a parameter. This shows that we can rely on just using @Inject
constructors for all classes that I create. Then I only need to add @Provides
methods for classes that are in the platform or in some other way not created by me.