Clean up!
Removed the CarModule, EngineModule and GarageModule classes since, as said in the last
commit, we do not need the @Provides methods if we annotate constructors in these classes
with the @Inject annotation, then these constructors will function as @Provides methods.
Also removed the CarComponent class since I never need to instantiate a Car class directly,
I only create Cars as part of a Garage object and so I only need to instantiate a Garage class.
Therefore I leave the GarageComponent class. I only need to have @Components to provide the end-point objects of the dependency-chain. Renamed the EngineModule class to IntegerModule since
it now only contains a @Provides Integer method.