Injecting external classes
In this commit I add a dependency on Context in the Engine class. Because of this and since Context
is an external class (I cannot add an @Inject constructor in it) I need to create a @Provides method
for the Context which I add in the IntegerModule class. Next, since I want the application context
and not just some empty Context object I need to make sure that the created @Provides method returns
the application context. I make sure that it does so by creating a constructor in IntegerModule that
accepts a Context and then using it. The way I use this constructor is that, when I create the Dagger
Component instead of calling the build() method directly on the builder returned by the builder()
method I instead first call the integerModule method and provide an instance of IntegerModule that
has been correctly created with the application context.